+++ src/Common/Time/Timer.cpp
@@ -9,7 +9,7 @@ double Timer::GetDelta()
 {
 	auto mCurrentTime = std::chrono::high_resolution_clock::now();
 	std::chrono::duration<double> elapsed = mCurrentTime - mStart;
-	double delta = std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count();
+	double delta = static_cast<double>(std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	mStart = mCurrentTime;
 
 	return delta;
